@@ -1,6 +1,6 @@
 package com.stefan.weathraw.ui.activity;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.Toast;
 
 import io.reactivex.disposables.CompositeDisposable;
