@@ -1,11 +1,11 @@
 package com.stefan.weathraw.ui.fragment;
 
 import android.appwidget.AppWidgetManager;
-import android.arch.lifecycle.Observer;
+import androidx.lifecycle.Observer;
 import android.content.ComponentName;
 import android.content.Intent;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import android.widget.Toast;
 
 import com.stefan.weathraw.ui.widget.BaseAppWidgetProvider;
