@@ -1,8 +1,8 @@
 package com.stefan.weathraw.cache;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.RoomDatabase;
+import androidx.room.Database;
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
 
 import com.stefan.weathraw.WeatherApplication;
 import com.stefan.weathraw.cache.dao.CityDao;
