@@ -1,10 +1,9 @@
 package com.stefan.weathraw.viewmodel;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
 
-import com.stefan.weathraw.model.ResponseMessage;
 import com.stefan.weathraw.utils.SingleLiveEvent;
 
 import io.reactivex.disposables.CompositeDisposable;
